MLK-22575-1: ASoC: fsl_easrc: configure slot according to pf memory size
The maximum prefilter memory size is 6144 entry, it is not allowed to
require memory size exceed this size.

When we calculate the available channel in the context processer, we
need to consider if the prefilter memory is not enough for the maxmum
channels, then we need to reduce the channels in this context processor,
move the left channel to another context processor.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
(cherry picked from commit 5a15d81a1c19c07846cc394d935524566bcfb298)