arm64: dts: ls1028a: fix dwc pci over smmu
In order for the dwc controller to work with SMMU it needs the
bootloader to fixup it's iommu-map property. In the current
implementation to bootloader will not perform the fixup if the
property is not already in the device tree with dummy fields.
Add it to fix DWC PCI over SMMU.

Signed-off-by: Laurentiu Tudor <laurentiu.tudor@nxp.com>