LF-841: drm/bridge: nwl-dsi: Fix mode_set
Currently, when setting a mode, if the new configuration matches the
existing configuration, the existing one is used instead. Also, the
drm_display_mode object was copied at this point, so if the new
configuration matched the existing one, the mode also wasn't copied,
causing the wrong mode to be used in DPI registers.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Reviewed-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>