LF-1115: arm64: dts: remove fixed-link node from external switch ports
Switch driver is now using Phylink instead of Phylib, so now external
switch ports (swp0-swp3) are managed in band, relies on SGMII AN between
the external PHY and internal SoC PHY.

In Linux, phylink normally manages both link on line side as well as link
between external PHY and SoC PCS. If configured as fixed link,
phylink re-configures SoC PCS to disable AN but does not touch the
external PHY configuration, the existence of an external
PHY is not known to it. So the effect is that PCS has now AN disabled and
the external PHY stays with AN on, based on u-boot configuration. These
settings are not compatible and SoC PCS and external PHY won't be able to
establish a link.

This patch removes fixed-link configuration from external switch ports
from fsl-ls1028a-rdb-dpdk.dts.

Earlier, fixed-link nodes were added to avoid HW issue ERR050089
(MDIO register issue). Now because external switch ports are managed
in-band and also phy-handle property is deleted, so there will be
no MDIO access for external PHYs and we can safely remove fixed-link
nodes.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
Acked-by: Alex Marginean <alexandru.marginean@nxp.com>