LF-17: crypto: caam/qi2: Fix support for TLS 1.0 record
When performing self test, the test failed with:
  dpaa2_caam dpseci.0: 40002502: DECO: desc idx 37: SGT Null Entry Error.
  alg: tls: encryption failed on test 0 for
        tls10-hmac-sha1-cbc-aes-caam-qi2: ret=5

The issue is wrong parameters are passed to sg_to_qm_sg_last when
creating the shared descriptor for TLS: the code was passing
the number of mapped SG entries instead of the length of
input data.

The issue was introduced because of a change of the
sg_to_qm_sg_last function [059d73eea640
(crypto: caam - use len instead of nents for bulding HW S/G table)]
and the patch to add TLS record was not modified

This patch fixes the parameter passed.

Fixes: 191ee3ffda9c (crypto: caam/qi2 - add support for TLS 1.0 record)
Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Horia Geantă <horia.geanta@nxp.com>
Signed-off-by: Horia Geantă <horia.geanta@nxp.com>